Move wind generation inside a folder
